--- a/compiler/src/iree/compiler/Dialect/HAL/Target/ROCM/CMakeLists.txt
+++ b/compiler/src/iree/compiler/Dialect/HAL/Target/ROCM/CMakeLists.txt
@@ -24,12 +24,15 @@ iree_cc_library(
     "ROCMTargetUtils.cpp"
   DEPS
     LLVMAMDGPUCodeGen
+    LLVMAnalysis
     LLVMCore
     LLVMIRReader
     LLVMLinker
     LLVMMC
+    LLVMPasses
     LLVMSupport
     LLVMTarget
+    LLVMTransformUtils
     LLVMipo
     MLIRAMDGPUDialect
     MLIRAMDGPUToROCDL