The Optimize Module performs series of optimizations on llvm module based on the sub-target (GFX9). E.g `llvm::CodeGenOptLevel::Aggressive`

The code is inspired by the code section in
https://github.com/openxla/iree/blob/main/compiler/plugins/target/CUDA/CUDATarget.cpp and
https://github.com/ROCmSoftwarePlatform/triton/blob/triton-mlir/lib/Target/HSACO/HSACOTranslation.cpp

**Before optimize module**

```
2023-11-27T18:28:29-08:00
Running ../iree-build/tools/iree-benchmark-module
Run on (64 X 7006.64 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x32)
  L1 Instruction 32 KiB (x32)
  L2 Unified 512 KiB (x32)
  L3 Unified 32768 KiB (x4)
Load Average: 0.54, 0.51, 0.63
***WARNING*** CPU scaling is enabled, the benchmark real time measurements may be noisy and will incur extra overhead.
***WARNING*** Library was built as DEBUG. Timings may be affected.
-------------------------------------------------------------------------------------------
Benchmark                                 Time             CPU   Iterations UserCounters...
-------------------------------------------------------------------------------------------
BM_matmul/process_time/real_time      48997 ms        48945 ms            1 items_per_second=0.0204094/s
```

**After optimize module**

```
2023-11-27T18:33:30-08:00
Running ../iree-build/tools/iree-benchmark-module
Run on (64 X 7006.64 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x32)
  L1 Instruction 32 KiB (x32)
  L2 Unified 512 KiB (x32)
  L3 Unified 32768 KiB (x4)
Load Average: 3.26, 3.42, 1.85
***WARNING*** CPU scaling is enabled, the benchmark real time measurements may be noisy and will incur extra overhead.
***WARNING*** Library was built as DEBUG. Timings may be affected.
-------------------------------------------------------------------------------------------
Benchmark                                 Time             CPU   Iterations UserCounters...
-------------------------------------------------------------------------------------------
BM_matmul/process_time/real_time      45699 ms        45668 ms            1 items_per_second=0.0218825/s
```