@erman-gurses 2023-11-22
> It would be good if we could find some way to test this. Also can you add a brief description to the PR describing what you're trying to accomplish?

I am checking again with some test cases that might be used. Also added a short description.