@erman-gurses 2023-11-28
> It would be good if we could find some way to test this. 

@qedawkins Please see below for this PR the performance differences before and after.

**Before optimize module**

2023-11-27T18:28:29-08:00
Running ../iree-build/tools/iree-benchmark-module
Run on (64 X 7006.64 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x32)
  L1 Instruction 32 KiB (x32)
  L2 Unified 512 KiB (x32)
  L3 Unified 32768 KiB (x4)
Load Average: 0.54, 0.51, 0.63
***WARNING*** CPU scaling is enabled, the benchmark real time measurements may be noisy and will incur extra overhead.
***WARNING*** Library was built as DEBUG. Timings may be affected.
-------------------------------------------------------------------------------------------
Benchmark                                 Time             CPU   Iterations UserCounters...
-------------------------------------------------------------------------------------------
BM_matmul/process_time/real_time      48997 ms        48945 ms            1 items_per_second=0.0204094/s

**After optimize module**

2023-11-27T18:33:30-08:00
Running ../iree-build/tools/iree-benchmark-module
Run on (64 X 7006.64 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x32)
  L1 Instruction 32 KiB (x32)
  L2 Unified 512 KiB (x32)
  L3 Unified 32768 KiB (x4)
Load Average: 3.26, 3.42, 1.85
***WARNING*** CPU scaling is enabled, the benchmark real time measurements may be noisy and will incur extra overhead.
***WARNING*** Library was built as DEBUG. Timings may be affected.
-------------------------------------------------------------------------------------------
Benchmark                                 Time             CPU   Iterations UserCounters...
-------------------------------------------------------------------------------------------
BM_matmul/process_time/real_time      45699 ms        45668 ms            1 items_per_second=0.0218825/s