@qedawkins 2023-11-28
Nice, that's good justification; @nirvedhmeshram seems to have more context on the llvm side, but is there any kind of reasonable e2e test worth adding here? Otherwise it would be good to log these performance numbers in the commit message as a record.