@qedawkins 2023-11-28
> Short of adding a benchmark CI there isn't much that we can do in terms of e2e test that measures performance. Also for the LLVM path itself, I don't know if we can add a lit test. This path is tested on all rocm tests, so it is tested, but can't really "test" for benchmark results afaik

Ok sorry for over-indexing on that then. I'm not too familiar with LLVM backend details so I will let others approve.