@ScottTodd 2023-11-28
We can at least add compilation correctness test coverage pretty easily. Execution correctness and performance test coverage will be more tricky.