@ScottTodd 2023-11-28
> > We can at least add compilation correctness test coverage pretty easily. Execution correctness and performance test coverage will be more tricky.
> 
> I think this path is effectively being tested by all rocm e2e tests...

Can you point to those tests? I don't think they exist right now, at least not here in IREE
I saw this thread this morning and planned on adding some today.

I see these ROCm tests right now:
```
iree/compiler/Dialect/HAL/Target/ROCM/test/smoketest.mlir.test
iree/compiler/Codegen/LLVMGPU/test/reduction_pipeline_rocm.mlir.test
iree/compiler/Codegen/LLVMGPU/test/reduction_pipeline_transform_rocm.mlir.test
experimental/regression_suite/tests/pregenerated/test_llama2.py::test_step_rdna3_rocm_stripped
```