@erman-gurses 2023-11-28
> Looks good to me, if the cloning can be addressed as a follow up.

Sure, I think it is mentioned in this issue https://github.com/openxla/iree/issues/15698